API Reference

Exported symbols:

@ga

KVector

Bivector

Trivector

Quadvector

codegen_expression

Main usage

SymbolicGA.@ga — Macro

@ga <sig> <flattening> <T> <ex>
@ga <sig> <flattening or T> <ex>
@ga <sig> <ex>

Generate Julia code which implements the computation of geometric elements from ex in an algebra defined by a signature sig (see SymbolicGA.Signature).

Supported syntax:

sig: Integer literal or tuple of 1, 2 or 3 integer literals corresponding to the number of positive, negative and degenerate basis vectors respectively, where unspecified integers default to zero.

flattening: Symbol literal.

T: Any arbitrary expression which evaluates to a type or to nothing.

ex: Any arbitrary expression that can be parsed algebraically.

See also: codegen_expression.

ex can be a single statement or a block, and uses a domain-specific language to facilitate the construction of algebraic expressions. ex is logically divided into two sections: a definition section, which defines bindings, and a final algebraic expression, which will be the object of the evaluation. It is processed in three phases:

A definition phase, in which bindings are defined with one or several statements for use in the subsequent phase;

An expansion phase, where identified bindings in the final algebraic expression are expanded. The defined bindings include the ones priorly defined and a set of built-in bindings.

An evaluation phase, in which the core algebraic expression is simplified and translated into a Julia expression.

Expression parsing

Binding definitions

All statements prior to the last can define new variables or functions with the following syntax and semantics:

Variables are either declared with <lhs::Symbol> = <rhs::Any> or with <lhs::Symbol>::<type>, the latter being expanded to <lhs> = <lhs>::<type>.

Functions are declared with a standard short or long form function definition <name>(<args...>) = <rhs> or function <name>(<args...>) <rhs> end, and are restricted to simple forms to encode simple semantics. The restrictions are as follows:

where clauses and output type annotations are not supported.

Function arguments must be untyped, e.g. f(x, y) is allowed but not f(x::Vector, y::Vector).

Function arguments must not be reassigned; it is assumed that any occurence of symbols declared as arguments will reference these arguments. For example, f(x, y) = x + y assumes that x + y actually means "perform + on the first and second function argument". Therefore, f(x, y) = (x = 2; x) + y will be likely to cause bugs. To alleviate this restriction, use codegen_expression with a suitable [SymbolicGA.VariableInfo] with function entries that contain specific calls to `:($(@arg <i>)).

Binding expansion

References and functions are expanded in a fairly straightforward copy-paste manner, where references are replaced with their right-hand side and function calls with their bodies with their arguments interpolated. Simple checks are put in place to allow for self-referencing bindings for references, such as x = x::T, leading to a single expansion of such a pattern in the corresponding expression subtree.

See SymbolicGA.VariableInfo for more information regarding the expansion of such variables and functions.

Algebraic evaluation

Type annotations may either:

Specify what type of geometric entity an input should be considered as, where components are then picked off with getcomponent.

Request the projection of an intermediate expression over one or multiple grades.

SymbolicGA.KVector — Type

KVector{K,T,D,N}

Geometric K-vector with eltype T with N elements in a geometric algebra of dimension D.

The constructors KVector{K,D}(elements...) and KVector{K,D}(elements::NTuple) will automatically infer T from the arguments and N from K and D.

Examples

julia> KVector{1,3}(1.0, 2.0, 3.0)
KVector{1, Float64, 3, 3}(1.0, 2.0, 3.0)

julia> KVector{2,3}(1.0, 2.0, 3.0)
Bivector{Float64, 3, 3}(1.0, 2.0, 3.0)

julia> KVector{4,4}(1.0)
Quadvector{Float64, 4, 1}(1.0)

SymbolicGA.Bivector — Type

Bivector{T,D,N}

Alias for KVector{2,T,D,N}

SymbolicGA.Trivector — Type

Trivector{T,D,N}

Alias for KVector{3,T,D,N}

SymbolicGA.Quadvector — Type

Quadvector{T,D,N}

Alias for KVector{4,T,D,N}

Expression generation

SymbolicGA.codegen_expression — Function

codegen_expression(sig, ex; flattening::Symbol = :nested, T = nothing, varinfo::Optional{VariableInfo} = nothing)

Parse ex as an algebraic expression and generate a Julia expression which represents the corresponding computation. sig can be a SymbolicGA.Signature or a signature integer, tuple or tuple expression adhering to semantics of @ga. See @ga for more information regarding the parsing and semantics applied to ex.

Parameters

flattening controls whether the components should be nested (:nested) or flattened (:flattened). In short, setting this option to :flattened always returns a single tuple of components, even multiple geometric entities are present in the output; while :nested will return a tuple of multiple elements if several geometric entities result from the computation.

T specifies what type to use when reconstructing geometric entities from tuple components with construct. If set to nothing with a :nested mode (default), then an appropriate KVector will be used depending on which type of geometric entity is returned; if multiple entities are present, a tuple of KVectors will be returned. With a :flattened mode, T will be set to :Tuple if unspecified.

varinfo is a user-provided SymbolicGA.VariableInfo which controls what expansions are carried out on the raw Julia expression before conversion to an algebraic expression.

SymbolicGA.VariableInfo — Type

VariableInfo(; refs = Dict{Symbol,Any}(), funcs = Dict{Symbol,Any}(), warn_override = true)

Structure holding information about bindings which either act as references (simple substitutions) or as functions, which can be called with arguments. This allows a small domain-specific language to be used when constructing algebraic expressions.

References are lhs::Symbol => rhs pairs where the left-hand side simply expands to the right-hand side during parsing. Right-hand sides which include lhs are supported, such that references of the form x = x::Vector are allowed, but will be expanded only once. Functions are name::Symbol => body pairs where rhs must refer to their arguments with Expr(:argument, <literal::Int>) expressions. Recursion is not supported and will lead to a StackOverflowError. See @arg.

Most built-in functions and symbols are implemented using this mechanism. If warn_override is set to true, overrides of such built-in functions will trigger a warning.

SymbolicGA.Signature — Type

Signature{P,N,D}

Signature of an Euclidean or pseudo-Euclidean space.

This signature encodes a space with a metric such that the first P basis vectors square to 1, the following N to -1 and the following D to 0. The metric evaluates to zero between two distinct basis vectors.

SymbolicGA.@arg — Macro

@arg <literal::Integer>

Convenience macro to construct expressions of the form Expr(:argument, i) used within function definitions for SymbolicGA.VariableInfo.

Interface methods

SymbolicGA.getcomponent — Function

getcomponent(collection)
getcomponent(collection, i)
getcomponent(collection, j, i)

Retrieve a number from a collection to be interpreted as the component of a geometric entity.

getcomponent(collection) which defaults to collection[] is used to extract the only component of a scalar or antiscalar.

getcomponent(collection, j, i) which defaults to collection[j][i] is used to extract the i-th component the j-th geometric entity for a collection of multiple geometric vectors.

getcomponent(collection, i) which defaults to collection[i] is used to extract the i-th component of a single geometric entity or a set of geometric entities backed by collection. In the case of a set of geometric entities, this is semantically equivalent to getcomponent(collection, j, i) where a single i is computed from the cumulated sum of past indices, i.e. as if components from consecutive entities were concatenated together. See @ga for more information regarding which mode is used depending on the syntax for grade extraction.

Most collections will not need to extend this method, but is exposed should the need arise for a tigher control over how input data is accessed.

SymbolicGA.construct — Function

construct(T, components::Tuple)

Construct an instance of T from a tuple of components.

Defaults to T(components).